# By Benjamin BONTEMPS
# Via Benjamin BONTEMPS
* 'master' of forge.lip6.fr:progandplay:
  repetition bugfix
  multiple games fully suported
  added lots of comments


Former-commit-id: 0cf4191565340351b1b8129ef5f7843fc237a6ae

+++ mods/Editor_Launcher/LuaUI/Widgets/launcher.lua
@@ -77,17 +77,17 @@ function InitializeMenu()
 end
 
 function Launch(game)
-	--if not VFS.FileExists("games/"..game.." (Prog&Play Level Editor).sdz") then
-		local modInfo = "return { game='PPLE', shortGame='PPLE', name='Prog & Play Level Editor', shortName='PPLE', mutator='official', version='0.1', description='A level editor for Prog & Play.', url='http://www.irit.fr/ProgAndPlay/index_en.php', modtype=0, depend= { \""..game.."\" },}"
+	if not VFS.FileExists("games/Prog & Play Level Editor for "..game..".sdz") then
+		local modInfo = "return { game='PPLE', shortGame='PPLE', name='Prog & Play Level Editor for "..game.."', shortName='PPLE', mutator='official', version='0.1', description='A level editor for Prog & Play.', url='http://www.irit.fr/ProgAndPlay/index_en.php', modtype=0, depend= { \""..game.."\" },}"
 		local file = io.open("games/editor/editor_files/ModInfo.lua", "w")
 		file:write(modInfo)
 		file:close()
 		VFS.CompressFolder("games/editor/editor_files")
-		os.rename("games/editor/editor_files.sdz", "games/Prog & Play Level Editor.sdz")
-	--end
+		os.rename("games/editor/editor_files.sdz", "games/Prog & Play Level Editor for "..game..".sdz")
+	end
 	local operations = {
 		["GAME"] = {
-			["Gametype"] = "Prog & Play Level Editor 0.1"
+			["Gametype"] = "Prog & Play Level Editor for "..game.." 0.1"
 		}
 	}
 	DoTheRestart("Editor.txt", operations)
